Move object around and refactor parseToGrafanaDashboardRequest and getScreenshotUrls
