+++ src/service/query/DashboardResponse.js
@@ -1,0 +1,20 @@
+class DashboardResponse {
+  constructor(imageUrl, grafanaChartLink, title) {
+    /**
+     * @type {string}
+     */
+    this.imageUrl = imageUrl;
+
+    /**
+     * @type {string}
+     */
+    this.grafanaChartLink = grafanaChartLink;
+
+    /**
+     * @type {string}
+     */
+    this.title = title;
+  }
+}
+
+exports.DashboardResponse = DashboardResponse;
